Working on demo scenes in testbed application

+++ testbed/common/HeightField.h
@@ -43,7 +43,7 @@ class HeightField : public openglframework::Mesh, public PhysicsObject {
         // -------------------- Attributes -------------------- //
 
         /// Height field data
-        float mHeightData[NB_POINTS_WIDTH * NB_POINTS_LENGTH];
+        int mHeightData[NB_POINTS_WIDTH * NB_POINTS_LENGTH];
 
         /// Previous transform (for interpolation)
         rp3d::Transform mPreviousTransform;
